New folder (2)/requirements.txt
fastapi
uvicorn
google-generativeai
sqlalchemy
jinja2
python-multipart
python-dotenv